Create README.md
I am looking for a partner with skills in app development to expand the functionality and features of my app. In exchange for your services I would be offering 50% share in the app. The first stage of the app has been built for IOS. The next stage is to create an Android version. I am currently working full time so this is not a full time project...... yet. For a better understanding of the concept you can visit www.tradiehubapp.com.au. 
If you are interested in hearing more please feel free to contact me via email. I am currently located in NYC